Vmstats should only be sent for full intervals
It was possible to see spikes or dips on restart of a service running
mondemand if you used the initial sample then the trigger to send
vmstats hit quickly because you'd have maybe 5 seconds of startup
stats.  Instead skip the first time sending stats, and only send on
the second time.  This will leave a gap in graphs on restart but will
mean you don't end up with spikes which dwarf the surrounding samples.