Set up flushing to allow for alternate logic.
I always have hated that the default mondemand-server config turned off
flushing in the library in order to replicate and do it's own.  So now
the flush functionality is pluggable, and the server just swaps out it's
own.  This also allows for consistent sending of vmstats (if enabled),
in all cases.